[BE] fix: 승인 대기중인 상태의 지원서만 삭제할 수 있도록 수정

+++ backend/src/main/java/com/devmatch/backend/domain/application/service/ApplicationService.java
@@ -54,6 +54,7 @@ public class ApplicationService {
       skillScores.add(score);
     }
 
+    // 멘토링 피드백: 세이브를 명시적으로 표현해주는 게 좋음
     application.getSkillScore().addAll(skillScores);
 
     return new ApplicationDetailResponseDto(applicationRepository.save(application));
@@ -104,6 +105,9 @@ public class ApplicationService {
   public void saveAnalysisResult(Long applicationId, AnalysisResult analysisResult) {
     Application application = getApplicationByApplicationId(applicationId);
 
+    // 멘토링 피드백: 종속 관계에 따라 구현을 어떤식으로 할지 고민.
+    // 두 서비스의 상위 서비스를 만들어서 관장하는 방식으로
+    // 상위 서비스의 이름은 보통 어플리케이션서비스라고 지음
     application.setAnalysisResult(analysisResult);
   }
 
@@ -111,6 +115,13 @@ public class ApplicationService {
   @Transactional
   public void deleteApplication(Long applicationId) {
     Application application = getApplicationByApplicationId(applicationId);
+    ApplicationStatus status = application.getStatus();
+
+    // 트러블 슈팅: 이미 승인 되어서 프로젝트 구성이 끝난 상태인데 지원서의 삭제가 가능함.
+    // -> 우리 프로젝트의 시나리오 상 승인 대기중인 지원서만 삭제할 수 있기 때문에 이외 상태는 예외 처리
+    if (!status.equals(ApplicationStatus.PENDING)) {
+      throw new IllegalArgumentException("승인 대기중인 지원서만 삭제할 수 있습니다. 현재 상태: " + status);
+    }
 
     applicationRepository.delete(application); // DB 에서 삭제
   }
